backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
openai==0.27.6
psycopg2==2.9.6